requirements.txt
flask
nltk
requests
openai
weaviate-client